Ported to UWP
Upgraded to VS 2017
