--- App And Page Lifecycle/AppLifecycleDemo.Universal/AppLifecycleDemo.Universal.Shared/AppLifecycleDemo.Universal.Shared.projitems
@@ -1,43 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<Project xmlns="http://schemas.microsoft.com/developer/msbuild/2003">
-  <PropertyGroup>
-    <MSBuildAllProjects>$(MSBuildAllProjects);$(MSBuildThisFileFullPath)</MSBuildAllProjects>
-    <HasSharedItems>true</HasSharedItems>
-    <SharedGUID>1b96210c-353a-4ac9-b7a3-4c11b47fed2b</SharedGUID>
-  </PropertyGroup>
-  <PropertyGroup Label="Configuration">
-    <Import_RootNamespace>AppLifecycleDemo.Universal</Import_RootNamespace>
-  </PropertyGroup>
-  <ItemGroup>
-    <ApplicationDefinition Include="$(MSBuildThisFileDirectory)App.xaml">
-      <SubType>Designer</SubType>
-    </ApplicationDefinition>
-    <Compile Include="$(MSBuildThisFileDirectory)App.xaml.cs">
-      <DependentUpon>App.xaml</DependentUpon>
-    </Compile>
-    <Compile Include="$(MSBuildThisFileDirectory)Services\AppLevelService.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)Views\ContentPage.xaml.cs">
-      <DependentUpon>ContentPage.xaml</DependentUpon>
-    </Compile>
-    <Compile Include="$(MSBuildThisFileDirectory)Views\MainPage.xaml.cs">
-      <DependentUpon>MainPage.xaml</DependentUpon>
-    </Compile>
-  </ItemGroup>
-  <ItemGroup>
-    <Content Include="$(MSBuildThisFileDirectory)Assets\FrogImages\FrogA.png" />
-    <Content Include="$(MSBuildThisFileDirectory)Assets\FrogImages\FrogB.png" />
-    <Content Include="$(MSBuildThisFileDirectory)Assets\FrogImages\FrogC.png" />
-    <Content Include="$(MSBuildThisFileDirectory)Assets\FrogImages\FrogD.png" />
-    <Content Include="$(MSBuildThisFileDirectory)Assets\FrogImages\FrogE.png" />
-  </ItemGroup>
-  <ItemGroup>
-    <Page Include="$(MSBuildThisFileDirectory)Views\ContentPage.xaml">
-      <SubType>Designer</SubType>
-      <Generator>MSBuild:Compile</Generator>
-    </Page>
-    <Page Include="$(MSBuildThisFileDirectory)Views\MainPage.xaml">
-      <SubType>Designer</SubType>
-      <Generator>MSBuild:Compile</Generator>
-    </Page>
-  </ItemGroup>
-</Project>
