--- App And Page Lifecycle/AppLifecycleDemo.Universal/AppLifecycleDemo.Universal.Windows/Package.appxmanifest
@@ -1,41 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<Package xmlns="http://schemas.microsoft.com/appx/2010/manifest" xmlns:m2="http://schemas.microsoft.com/appx/2013/manifest">
-
-  <Identity Name="ebd0aa06-223b-4b77-be72-7b4ee08cc4f8"
-            Publisher="CN=Danny"
-            Version="1.0.0.0" />
-
-  <Properties>
-    <DisplayName>AppLifecycleDemo.Universal.Windows</DisplayName>
-    <PublisherDisplayName>Danny</PublisherDisplayName>
-    <Logo>Assets\StoreLogo.png</Logo>
-  </Properties>
-
-  <Prerequisites>
-    <OSMinVersion>6.3.0</OSMinVersion>
-    <OSMaxVersionTested>6.3.0</OSMaxVersionTested>
-  </Prerequisites>
-
-  <Resources>
-    <Resource Language="x-generate"/>
-  </Resources>
-
-  <Applications>
-    <Application Id="App"
-        Executable="$targetnametoken$.exe"
-        EntryPoint="AppLifecycleDemo.Universal.Windows.App">
-        <m2:VisualElements
-            DisplayName="AppLifecycleDemo.Universal.Windows"
-            Square150x150Logo="Assets\Logo.png"
-            Square30x30Logo="Assets\SmallLogo.png"
-            Description="AppLifecycleDemo.Universal.Windows"
-            ForegroundText="light"
-            BackgroundColor="#464646">
-            <m2:SplashScreen Image="Assets\SplashScreen.png" />
-        </m2:VisualElements>
-    </Application>
-  </Applications>
-  <Capabilities>
-    <Capability Name="internetClient" />
-  </Capabilities>
-</Package>
