--- App And Page Lifecycle/AppLifecycleDemo.Universal/AppLifecycleDemo.Universal.WindowsPhone/Package.appxmanifest
@@ -1,44 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<Package xmlns="http://schemas.microsoft.com/appx/2010/manifest" xmlns:m2="http://schemas.microsoft.com/appx/2013/manifest" xmlns:m3="http://schemas.microsoft.com/appx/2014/manifest" xmlns:mp="http://schemas.microsoft.com/appx/2014/phone/manifest">
-
-  <Identity Name="ee05774b-75fe-4f4b-be62-317c9d54a48c"
-            Publisher="CN=Danny"
-            Version="1.0.0.0" />
-
-  <mp:PhoneIdentity PhoneProductId="ee05774b-75fe-4f4b-be62-317c9d54a48c" PhonePublisherId="00000000-0000-0000-0000-000000000000"/>
-
-  <Properties>
-    <DisplayName>AppLifecycleDemo.Universal.WindowsPhone</DisplayName>
-    <PublisherDisplayName>Danny</PublisherDisplayName>
-    <Logo>Assets\StoreLogo.png</Logo>
-  </Properties>
-
-  <Prerequisites>
-    <OSMinVersion>6.3.1</OSMinVersion>
-    <OSMaxVersionTested>6.3.1</OSMaxVersionTested>
-  </Prerequisites>
-  
-  <Resources>
-    <Resource Language="x-generate"/>
-  </Resources>
-
-  <Applications>
-    <Application Id="App"
-        Executable="$targetnametoken$.exe"
-        EntryPoint="AppLifecycleDemo.Universal.WindowsPhone.App">
-        <m3:VisualElements
-            DisplayName="AppLifecycleDemo.Universal.WindowsPhone"
-            Square150x150Logo="Assets\Logo.png"
-            Square44x44Logo="Assets\SmallLogo.png"
-            Description="AppLifecycleDemo.Universal.WindowsPhone"
-            ForegroundText="light"
-            BackgroundColor="transparent">
-            <m3:DefaultTile Wide310x150Logo="Assets\WideLogo.png" Square71x71Logo="Assets\Square71x71Logo.png"/>
-            <m3:SplashScreen Image="Assets\SplashScreen.png"/>
-        </m3:VisualElements>
-    </Application>
-  </Applications>
-  <Capabilities>
-    <Capability Name="internetClientServer" />
-  </Capabilities>
-</Package>
