+++ App And Page Lifecycle/AppLifecycleDemo.Uwp/App.xaml
@@ -1,7 +1,8 @@
 ﻿<Application
-    x:Class="AppLifecycleDemo.Universal.App"
+    x:Class="AppLifecycleDemo.Uwp.App"
     xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
     xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-    xmlns:local="using:AppLifecycleDemo.Universal">
+    xmlns:local="using:AppLifecycleDemo.Uwp"
+    RequestedTheme="Light">
 
 </Application>
