+++ App And Page Lifecycle/AppLifecycleDemo.Uwp/Assets/FrogImages/FrogA.png
