+++ App And Page Lifecycle/AppLifecycleDemo.Uwp/Assets/FrogImages/FrogB.png
