+++ App And Page Lifecycle/AppLifecycleDemo.Uwp/Assets/FrogImages/FrogC.png
