+++ App And Page Lifecycle/AppLifecycleDemo.Uwp/Assets/FrogImages/FrogD.png
