+++ App And Page Lifecycle/AppLifecycleDemo.Uwp/Assets/FrogImages/FrogE.png
