+++ App And Page Lifecycle/AppLifecycleDemo.Uwp/Views/ContentPage.xaml
@@ -1,7 +1,7 @@
-﻿<Page x:Class="AppLifecycleDemo.Universal.Views.ContentPage"
+﻿<Page x:Class="AppLifecycleDemo.Uwp.Views.ContentPage"
 	  xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
 	  xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-	  xmlns:local="using:AppLifecycleDemo.Universal.Views"
+	  xmlns:local="using:AppLifecycleDemo.Uwp.Views"
 	  xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
 	  xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
 	  mc:Ignorable="d">
