+++ App And Page Lifecycle/AppLifecycleDemo.Uwp/Views/MainPage.xaml
@@ -1,16 +1,30 @@
-﻿<Page x:Class="AppLifecycleDemo.Universal.Views.MainPage"
+﻿<Page x:Class="AppLifecycleDemo.Uwp.Views.MainPage"
 	  xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
 	  xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-	  xmlns:local="using:AppLifecycleDemo.Universal.Views"
+	  xmlns:local="using:AppLifecycleDemo.Uwp.Views"
 	  xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
 	  xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
 	  mc:Ignorable="d">
+
+    <Page.Resources>
+        <Style TargetType="Button" >
+            <Setter Property="Margin" Value="0,10"/>
+            <Setter Property="Width" Value="50"/>
+            <Setter Property="Height" Value="50"/>
+            <Setter Property="FontSize" Value="25"/>
+        </Style>
+    </Page.Resources>
 
 
     <Grid x:Name="RootGrid"
               Background="{ThemeResource ApplicationPageBackgroundThemeBrush}"
               Margin="0,20">
+        <Grid.ColumnDefinitions>
+            <ColumnDefinition/>
+            <ColumnDefinition/>
+        </Grid.ColumnDefinitions>
         <StackPanel x:Name="SecondsSinceLaunchSection"
+                    Grid.Column="0"
                     Margin="12,0">
             <TextBlock x:Name="SecondsSinceLaunchHeaderTextBlock"
                        Text="Seconds Since Launch"
@@ -23,7 +37,9 @@
             <TextBlock Text="{Binding Path=SecondsAppInUse}"
 					   Style="{StaticResource HeaderTextBlockStyle}" />
         </StackPanel>
-        <StackPanel x:Name="ActionSection"  Margin="12,0">
+        <StackPanel x:Name="ActionSection"
+                    Grid.Column="1"
+                    Margin="12,0">
             <TextBlock Text="App Lifecycle Demo"
 					   Style="{StaticResource HeaderTextBlockStyle}" />
             <Button x:Name="PartA"
