+++ App And Page Lifecycle/AppLifecycleDemo.Uwp/Views/MainPage.xaml.cs
@@ -1,32 +1,15 @@
 ﻿using Windows.UI.Xaml;
 using Windows.UI.Xaml.Controls;
-using AppLifecycleDemo.Universal.Services;
+using AppLifecycleDemo.Uwp.Services;
 
-namespace AppLifecycleDemo.Universal.Views
+namespace AppLifecycleDemo.Uwp.Views
 {
-    /// <summary>
-    /// An empty page that can be used on its own or navigated to within a Frame.
-    /// </summary>
     public sealed partial class MainPage : Page
     {
         public MainPage()
         {
             this.InitializeComponent();
             DataContext = AppLevelService.Instance;
-
-#if WINDOWS_APP
-            RootGrid.ColumnDefinitions.Add(new ColumnDefinition());
-            RootGrid.ColumnDefinitions.Add(new ColumnDefinition());
-            Grid.SetColumn(ActionSection, 0);
-            Grid.SetColumn(SecondsSinceLaunchSection, 1);
-#elif WINDOWS_PHONE_APP
-            RootGrid.RowDefinitions.Add(new RowDefinition() { Height = GridLength.Auto });
-            RootGrid.RowDefinitions.Add(new RowDefinition());
-            Grid.SetRow(ActionSection, 1);
-            Grid.SetRow(SecondsSinceLaunchSection, 0);
-            SecondsSinceLaunchHeaderTextBlock.FontSize = 40;
-            SecondsInUseHeaderTextBlock.FontSize = 40;
-#endif
         }
 
         private void PartA_Click( object sender, RoutedEventArgs routedEventArgs )
